+++ blocklets/image-bin/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.11.115 (April 28, 2025)
+
+- chore: update deps
+- chore: polish lottie json files import logic
+
 ## 0.11.114 (April 18, 2025)
 
 - chore: update deps